finished avl tree(?) added some tests

+++ src/main/java/org/example/tree/DataAdapter.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface DataAdapter {
     boolean addOrUpdate(Person person);
-    boolean delete(int inn);
-    Person get(int inn);
+    boolean delete(long inn);
+    Person get(long inn);
     List<Person> get();
 }
